requirements.txt
fastapi
uvicorn[standard]
pydantic
youtube-transcript-api
openai
python-dotenv
